Refactor dagster jobs into a single module
After using Dagster for a while it seems I rarely need more than a
single Dagster job module that contain my configured graphs.
