+++ src/main.py
@@ -1,5 +1,5 @@
 """Example of how to run a Dagster op from normal Python script."""
-from jobs.cereal_jobs import complex_job
+from jobs import complex_job
 
 if __name__ == "__main__":
     result = complex_job.execute_in_process()
